add loadhook function so i can be compiled on different os
