+++ plugin_export_notwin.go
@@ -1,0 +1,5 @@
+// +build !windows
+
+package vst2
+
+func loadHook() {}
